requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==39.0.1
lxml==4.9.2
ncclient==0.6.13
paramiko==3.0.0
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
six==1.16.0
